Refactoring and documentation : - Moved away chunkers and filters from the main __init__ for easiness of development - Changed documentation of API - Documentation of plugins design - Updated examples - Contribution Guidelines
