+++ docs/Nemo.chunker.rst
@@ -93,6 +93,7 @@ Available chunkers
 .. automethod:: flask.ext.nemo.Nemo.line_chunker
 .. automethod:: flask.ext.nemo.Nemo.scheme_chunker
 .. automethod:: flask.ext.nemo.Nemo.level_chunker
+.. automethod:: flask.ext.nemo.Nemo.level_grouper
 
 PrevNext
 ########
