+++ examples/translations/r_double.html
@@ -1,4 +1,4 @@
-{% extends "container.html" %}
+{% extends "main::container.html" %}
 
 {% block article %}
 <div class="row">
