+++ flask_nemo/common.py
@@ -1,5 +1,12 @@
 import os.path as op
 from collections import OrderedDict
+import re
+from functools import reduce
+
+
+""" Regular expression to match common literature namespace
+"""
+regMatch = re.compile("^[a-z]{3}Lit$")
 
 
 def resource_qualifier(resource):
@@ -18,3 +25,41 @@ ASSETS_STRUCTURE = {
     "css": OrderedDict(),
     "static": OrderedDict()
 }
+
+
+def join_or_single(start, end):
+    """ Join passages range. If they are the same, return a single part of the range
+
+    :param start: Start of the passage range
+    :param end: End of the passage range
+    :return: Finale Passage Chunk Identifier
+    """
+    if start == end:
+        return start
+    else:
+        return "{}-{}".format(
+            start,
+            end
+        )
+
+
+def getFromDict(dataDict, keyList):
+    """Retrieves and creates when necessary a dictionary in nested dictionaries
+
+    :param dataDict: a dictionary
+    :param keyList: list of keys
+    :return: target dictionary
+    """
+    return reduce(create_hierarchy, keyList, dataDict)
+
+
+def create_hierarchy(hierarchy, level):
+    """Create an OrderedDict
+
+    :param hierarchy: a dictionary
+    :param level: single key
+    :return: deeper dictionary
+    """
+    if level not in hierarchy:
+        hierarchy[level] = OrderedDict()
+    return hierarchy[level]
