+++ tests/test_filters.py
@@ -6,9 +6,11 @@ Test file for filters.
 
 from flask import Flask
 from flask_nemo import Nemo
-from mock import patch, call, Mock
+from flask_nemo.filters import f_active_link, f_collection_i18n, f_formatting_passage_reference, f_group_texts, \
+    f_hierarchical_passages, f_i18n_citation_type, f_i18n_iso, f_is_str, f_order_author,\
+    f_order_text_edition_translation
 import MyCapytain
-from .resources import NemoResource, RequestPatch, RequestPatchChained, create_test_app
+from .resources import NemoResource
 
 
 class TestFilters(NemoResource):
@@ -30,7 +32,7 @@ class TestFilters(NemoResource):
         d.metadata["groupname"]["eng"] = "Augustine"
         e.metadata["groupname"]["eng"] = "Suetone"
         f.metadata["groupname"]["eng"] = None
-        sort = Nemo.f_order_author([a, b, c, d, e, f])
+        sort = f_order_author([a, b, c, d, e, f])
         self.assertEqual(
             sort,
             [d, c, a, e, b, f]
@@ -39,9 +41,9 @@ class TestFilters(NemoResource):
     def test_f_active_link(self):
         """ Test checking if something is in the path
         """
-        wrong = Nemo.f_active_link("phi003", {"collection": "latinLit"})
-        wrong2 = Nemo.f_active_link("phi003", {"collection": "latinLit", "textgroup": "phi0003"})
-        right = Nemo.f_active_link("phi003", {"collection": "latinLit", "textgroup": "phi0003", "work": "phi003"})
+        wrong = f_active_link("phi003", {"collection": "latinLit"})
+        wrong2 = f_active_link("phi003", {"collection": "latinLit", "textgroup": "phi0003"})
+        right = f_active_link("phi003", {"collection": "latinLit", "textgroup": "phi0003", "work": "phi003"})
         self.assertEqual(wrong, "")
         self.assertEqual(wrong2, "")
         self.assertEqual(right, "active")
@@ -49,37 +51,37 @@ class TestFilters(NemoResource):
     def test_f_collection_i18n(self):
         """ Test internationalization of collection identifier
         """
-        self.assertEqual(Nemo.f_collection_i18n("latinLit"), "Latin")
-        self.assertEqual(Nemo.f_collection_i18n("greekLit"), "Ancient Greek")
-        self.assertEqual(Nemo.f_collection_i18n("freLit"), "freLit")
+        self.assertEqual(f_collection_i18n("latinLit"), "Latin")
+        self.assertEqual(f_collection_i18n("greekLit"), "Ancient Greek")
+        self.assertEqual(f_collection_i18n("freLit"), "freLit")
 
     def test_f_formatting_passage_reference(self):
         """ Test split of passage range identifier
         """
-        self.assertEqual(Nemo.f_formatting_passage_reference("1.1-1.2"), "1.1")
-        self.assertEqual(Nemo.f_formatting_passage_reference("1.1"), "1.1")
+        self.assertEqual(f_formatting_passage_reference("1.1-1.2"), "1.1")
+        self.assertEqual(f_formatting_passage_reference("1.1"), "1.1")
 
     def test_f_i18n_iso(self):
         """ Test split of passage range identifier
         """
-        self.assertEqual(Nemo.f_i18n_iso("eng"), "English")
-        self.assertEqual(Nemo.f_i18n_iso("eng", "fre"), "anglais")
-        self.assertEqual(Nemo.f_i18n_iso("eng", "ger"), "English")
+        self.assertEqual(f_i18n_iso("eng"), "English")
+        self.assertEqual(f_i18n_iso("eng", "fre"), "anglais")
+        self.assertEqual(f_i18n_iso("eng", "ger"), "English")
 
     def test_f_order_text_edition_translation(self):
         """ Check the reordering filter
         """
         Text = MyCapytain.resources.inventory.Text
         a, b, c, d = Text(subtype="Translation"), Text(subtype="Edition"), Text(subtype="Edition"), Text(subtype="Translation")
-        self.assertEqual(Nemo.f_order_text_edition_translation([a,b,c,d]), [b, c, a, d])
+        self.assertEqual(f_order_text_edition_translation([a,b,c,d]), [b, c, a, d])
 
     def test_register_filter(self):
         app = Flask(__name__)
         self.nemo = Nemo(app=app)
         self.nemo.register_filters()
-        self.assertEqual(self.nemo.app.jinja_env.filters["formatting_passage_reference"], Nemo.f_formatting_passage_reference)
-        self.assertEqual(self.nemo.app.jinja_env.filters["collection_i18n"], Nemo.f_collection_i18n)
-        self.assertEqual(self.nemo.app.jinja_env.filters["active_link"], Nemo.f_active_link)
+        self.assertEqual(self.nemo.app.jinja_env.filters["formatting_passage_reference"], f_formatting_passage_reference)
+        self.assertEqual(self.nemo.app.jinja_env.filters["collection_i18n"], f_collection_i18n)
+        self.assertEqual(self.nemo.app.jinja_env.filters["active_link"], f_active_link)
 
     def test_f_hierarchical_passages(self):
         """Test for normal, simple passage hierarchical conversion
@@ -91,7 +93,7 @@ class TestFilters(NemoResource):
         citation_book = MyCapytain.common.reference.Citation(name="book", child=citation_poem)
         text = MyCapytain.resources.inventory.Text()
         text.citation = citation_book
-        converted = Nemo.f_hierarchical_passages(reffs, text)
+        converted = f_hierarchical_passages(reffs, text)
         self.assertEqual(converted["%book|1%"]["%poem|5%"]["Line 8"], "1.5.8")
         self.assertEqual(converted["%book|1%"]["%poem|5%"]["Line 9"], "1.5.9")
         self.assertEqual(converted["%book|1%"]["%poem|6%"]["Line 7"], "1.6.8")
@@ -112,7 +114,7 @@ class TestFilters(NemoResource):
         citation_book = MyCapytain.common.reference.Citation(name="book", child=citation_poem)
         text = MyCapytain.resources.inventory.Text()
         text.citation = citation_book
-        converted = Nemo.f_hierarchical_passages(reffs, text)
+        converted = f_hierarchical_passages(reffs, text)
         self.assertEqual(converted["%book|1%"]["%poem|5%"]["Line 8"], "1.5.8-1.5.9")
         self.assertEqual(converted["%book|1%"]["%poem|5%"]["Line 9"], "1.5.9-1.5.15")
         self.assertEqual(converted["%book|1%"]["%poem|6%"]["Line 7"], "1.6.8-2.9.16")
@@ -126,8 +128,8 @@ class TestFilters(NemoResource):
     def test_f_is_str(self):
         """ Test string
         """
-        self.assertEqual(Nemo.f_is_str("h"), True)
-        self.assertEqual(Nemo.f_is_str([]), False)
+        self.assertEqual(f_is_str("h"), True)
+        self.assertEqual(f_is_str([]), False)
 
     def test_f_i18n_citation_type(self):
-        self.assertEqual(Nemo.f_i18n_citation_type("%book|1%"), "Book 1")
+        self.assertEqual(f_i18n_citation_type("%book|1%"), "Book 1")
